refactor: migrate from ffi-napi to process-based native tools architecture
- Replace ffi-napi/DLL approach with process-based execution matching macOS architecture
- Implement three separate C# executables (window-detector, keyboard-simulator, text-field-detector)
- Update windows-platform-tools.ts to use child_process.exec instead of FFI
- Remove koffi and ffi-napi dependencies from package.json
- Update build scripts to compile individual C# projects
- Add proper .gitignore entries for Windows build artifacts
- Update documentation to reflect new architecture

This change provides:
- Unified architecture between macOS and Windows
- Simpler debugging and maintenance
- Elimination of complex FFI dependencies
- Better process isolation and security
