+++ kill-electron.ps1
@@ -1,0 +1,2 @@
+Stop-Process -Name "electron" -Force -ErrorAction SilentlyContinue
+Write-Host "All Electron processes have been stopped"
