+++ prompt-line.sln
@@ -1,0 +1,24 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.2.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowDetector", "native-win\WindowDetector.csproj", "{BBAF8351-212E-3120-DBA3-5568962D03F8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BBAF8351-212E-3120-DBA3-5568962D03F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BBAF8351-212E-3120-DBA3-5568962D03F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BBAF8351-212E-3120-DBA3-5568962D03F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BBAF8351-212E-3120-DBA3-5568962D03F8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {333A2838-D395-4793-821F-2259F39533D7}
+	EndGlobalSection
+EndGlobal
