--- src/native-tools/WindowDetector.runtimeconfig.json
@@ -1,21 +1,0 @@
-{
-  "runtimeOptions": {
-    "tfm": "net8.0",
-    "rollForward": "LatestMinor",
-    "frameworks": [
-      {
-        "name": "Microsoft.NETCore.App",
-        "version": "8.0.0"
-      },
-      {
-        "name": "Microsoft.WindowsDesktop.App",
-        "version": "8.0.0"
-      }
-    ],
-    "configProperties": {
-      "System.Reflection.Metadata.MetadataUpdater.IsSupported": false,
-      "System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization": true,
-      "CSWINRT_USE_WINDOWS_UI_XAML_PROJECTIONS": false
-    }
-  }
-}
