--- src/native-tools/keyboard-simulator.exe
@@ -1,1 +1,0 @@
-# Windows native tool placeholder
