--- src/native-tools/text-field-detector.exe
@@ -1,1 +1,0 @@
-# Windows native tool placeholder
