--- src/native-tools/window-detector.exe
@@ -1,1 +1,0 @@
-# Windows native tool placeholder
